+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.15.0] - 2026-09-22
+
 ### Added
 
 - Require the `downloadNotificationRequests` License Manager resource in addition to admin role to access Download Requests, Process Unsent and Delete Request APIs.